# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface VideoHighlight {
  dish: string;
  tagline: string;
  /**
     * @minItems 3
     * @maxItems 3
     */
  descriptors: string[];
  accentColor: string;
}

export interface Venue {
  id: string;
  name: string;
  category: string;
  rating: number;
  reviewCount: number;
  address: string;
  city: string;
  neighborhood: string;
  phone: string;
  website: string;
  hours: string;
  description: string;
  tags: string[];
  featured: boolean;
  verified: boolean;
  priceRange: string;
  color: string;
  initials: string;
  hasVideo: boolean;
  video?: VideoHighlight | null;
  lat: number;
  lng: number;
}

export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface FollowsResponse {
  following: string[];
}

export interface ToggleFollowResponse {
  following: boolean;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};