# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /follows:
    get:
      tags: [Follows]
      operationId: getMyFollows
      summary: Get IDs of venues the current user follows
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of followed venue IDs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowsResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /follows/{venueId}:
    post:
      tags: [Follows]
      operationId: toggleVenueFollow
      summary: Toggle follow/unfollow a venue
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: venueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Follow state after toggle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToggleFollowResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    FollowsResponse:
      type: object
      required: [following]
      properties:
        following:
          type: array
          items:
            type: string
    ToggleFollowResponse:
      type: object
      required: [following]
      properties:
        following:
          type: boolean